Inline abstract class
